Ignore single space around code span
According to https://spec.commonmark.org/0.29/#code-spans :

> If the resulting string both begins and ends with a space character,
but does not consist entirely of space characters, a single space
character is removed from the front and back. This allows you to include
code that begins or ends with backtick characters, which must be
separated by whitespace from the opening or closing backtick strings.